hotfix: changed minheight to allow lists to grow in size
I still want a min-height of 350px, but at the moment that's conflicting with min-h-min. I'm not sure why the section height won't grow with the content, but I will look into that when I'm not tired. For now, this is a fix that will allow me to use the app tomorrow.